refactor: combine two hooks
