+++ packages/shared/src/components/squads/SquadTourPopup.tsx
@@ -4,22 +4,17 @@ import InteractivePopup, {
   InteractivePopupPosition,
 } from '../tooltips/InteractivePopup';
 import SquadTour from './SquadTour';
-import { useSquadTourClose } from '../../hooks/useSquadTourClose';
+import { useSquadTour } from '../../hooks/useSquadTour';
 
 interface SquadTourPopupProps {
   onClose: React.EventHandler<MouseEvent>;
 }
 
 function SquadTourPopup({ onClose }: SquadTourPopupProps): ReactElement {
-  const sharePostTutorial = useTutorial({
-    key: TutorialKey.SHARE_SQUAD_POST,
-  });
-
-  const [onSquadTourClose] = useSquadTourClose();
+  const { onCloseTour } = useSquadTour();
   const onPopupClose = (event: MouseEvent) => {
-    onTourIndexChange(-1);
     onClose(event);
-    onSquadTourClose();
+    onCloseTour();
   };
 
   return (
