+++ packages/shared/src/hooks/useSquadTour.ts
@@ -1,5 +1,6 @@
 import { useQuery, useQueryClient } from 'react-query';
 import { useMemo } from 'react';
+import { TutorialKey, useTutorial } from './useTutorial';
 
 const SQUAD_TOUR_KEY = 'squadTourKey';
 
@@ -9,6 +10,7 @@ interface SquadTourData {
 
 interface UseSquadTour extends SquadTourData {
   onTourIndexChange: (value: number) => void;
+  onCloseTour: () => void;
 }
 
 export const useSquadTour = (): UseSquadTour => {
@@ -18,6 +20,9 @@ export const useSquadTour = (): UseSquadTour => {
     () => client.getQueryData(SQUAD_TOUR_KEY),
     { initialData: { tourIndex: -1 } },
   );
+  const sharePostTutorial = useTutorial({
+    key: TutorialKey.ShareSquadPost,
+  });
 
   const onTourIndexChange = (tourIndex: number) =>
     client.setQueryData<SquadTourData>(SQUAD_TOUR_KEY, (value) => ({
@@ -25,5 +30,15 @@ export const useSquadTour = (): UseSquadTour => {
       tourIndex,
     }));
 
-  return useMemo(() => ({ ...data, onTourIndexChange }), [data]);
+  const onClose = () => {
+    onTourIndexChange(-1);
+    if (!sharePostTutorial.isCompleted) {
+      sharePostTutorial.activate();
+    }
+  };
+
+  return useMemo(
+    () => ({ ...data, onCloseTour: onClose, onTourIndexChange }),
+    [data, sharePostTutorial],
+  );
 };
