--- packages/shared/src/hooks/useSquadTourClose.ts
@@ -1,15 +1,0 @@
-import { TutorialKey, useTutorial } from './useTutorial';
-
-export function useSquadTourClose(): [() => void] {
-  const sharePostTutorial = useTutorial({
-    key: TutorialKey.ShareSquadPost,
-  });
-
-  const onClose = () => {
-    if (!sharePostTutorial.isCompleted) {
-      sharePostTutorial.activate();
-    }
-  };
-
-  return [onClose];
-}
